+++ Makefile
@@ -14,7 +14,7 @@ portable_zip: build
 		mkdir portable_iblinter
 		mkdir portable_iblinter/lib
 		mkdir portable_iblinter/bin
-		cp -f .build/release/main portable_iblinter/bin/iblinter
+		cp -f .build/release/iblinter portable_iblinter/bin/iblinter
 		cp -rf $(C_LIB_DIRS) $(SWIFT_LIB_FILES) "portable_iblinter/lib"
 		cp -f LICENSE portable_iblinter
 		cd portable_iblinter
@@ -24,7 +24,7 @@ portable_zip: build
 install: build
 		mkdir -p "$(PREFIX)/bin"
 		mkdir -p "$(PREFIX)/lib/iblinter"
-		cp -f ".build/release/main" "$(PREFIX)/bin/iblinter"
+		cp -f ".build/release/iblinter" "$(PREFIX)/bin/iblinter"
 		cp -rf $(C_LIB_DIRS) $(SWIFT_LIB_FILES) "$(PREFIX)/lib/iblinter"
 
 current_version:
