+++ Sources/IBLinter/main.swift
@@ -1,4 +1,4 @@
-import IBLinter
+import IBLinterFrontend
 
 let app = IBLinter()
 app.run()
